pyproject.toml
[project]
name = "agentic-jso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chardet<6",
    "deepagents>=0.4.10",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "numpy<2",
    "py2pdf>=0.1.3",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "pyresparser>=1.0.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "resume-parser>=0.8.4",
    "setuptools<81",
    "spacy>=2.3.9,<2.4",
    "tavily-python>=0.7.23",
    "uvicorn>=0.41.0",
]
